Miniproyecto2_SSD/dl-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
numpy==1.26.2
pillow==10.1.0
python-multipart==0.0.6
